# Greenovation-backend-main/docs/openapiGreenovation.yaml
openapi: 3.0.3
info:
  version: 1.0.0
  title: Greenovation
  description: "This is the Open API documentation for the Greenovation API"
servers:
  - url: 'https://quason.tech/'
  - url: 'http://nattech.fib.upc.edu:40370/'
  - url: 'http://localhost:3000/'
paths:
...
  /dadesObertes:
    get:
      tags:
        - Dades Obertes
      summary: Returns all results from DadesObertes having filtered them according to given parameters. For more details consult https://dev.socrata.com/foundry/analisi.transparenciacatalunya.cat/j6ii-t3w2
      operationId: getEdificisDadesObertes
      parameters:
        - name: num_cas
          in: query
          required: false
          schema:
            type: string
        - name: adre_a
          in: query
          required: false
          schema:
            type: string
        - name: numero
          in: query
          required: false
          schema:
            type: string
        - name: escala
          in: query
          required: false
          schema:
            type: string
        - name: pis
          in: query
          required: false
          schema:
            type: string
        - name: porta
          in: query
          required: false
          schema:
            type: string
        - name: codi_postal
          in: query
          required: false
          schema:
            type: string
        - name: poblacio
          in: query
          required: false
          schema:
            type: string
        - name: comarca
          in: query
          required: false
          schema:
            type: string
        - name: nom_provincia
          in: query
          required: false
          schema:
            type: string
        - name: codi_poblacio
          in: query
          required: false
          schema:
            type: string
        - name: codi_comarca
          in: query
          required: false
          schema:
            type: string
        - name: codi_provincia
          in: query
          required: false
          schema:
            type: string
        - name: referencia_cadastral
          in: query
          required: false
          schema:
            type: string
        - name: zona_climatica
          in: query
          required: false
          schema:
            type: string
        - name: metres_cadastre
          in: query
          required: false
          schema:
            type: number
        - name: us_edifici
          in: query
          required: false
          schema:
            type: string
        - name: qualificaci_de_consum_d
          in: query
          required: false
          schema:
            type: string
        - name: energia_prim_ria_no_renovable
          in: query
          required: false
          schema:
            type: number
        - name: qualificacio_d_emissions
          in: query
          required: false
          schema:
            type: string
        - name: emissions_de_co2
          in: query
          required: false
          schema:
            type: number
        - name: consum_d_energia_final
          in: query
          required: false
          schema:
            type: number
        - name: cost_anual_aproximat_d_energia
          in: query
          required: false
          schema:
            type: number
        - name: vehicle_electric
          in: query
          required: false
          schema:
            type: string
        - name: solar_termica
          in: query
          required: false
          schema:
            type: string
        - name: solar_fotovoltaica
          in: query
          required: false
          schema:
            type: string
        - name: sistema_biomassa
          in: query
          required: false
          schema:
            type: string
        - name: xarxa_districte
          in: query
          required: false
          schema:
            type: string
        - name: energia_geotermica
          in: query
          required: false
          schema:
            type: string
        - name: eina_de_certificacio
          in: query
          required: false
          schema:
            type: string
        - name: valor_aillaments
          in: query
          required: false
          schema:
            type: number
        - name: valor_finestres
          in: query
          required: false
          schema:
            type: number
        - name: motiu_de_la_certificacio
          in: query
          required: false
          schema:
            type: string
        - name: valor_aillaments_cte
          in: query
          required: false
          schema:
            type: number
        - name: valor_finestres_cte
          in: query
          required: false
          schema:
            type: number
        - name: utm_x
          in: query
          required: false
          schema:
            type: number
        - name: utm_y
          in: query
          required: false
          schema:
            type: number
        - name: normativa_construcci
          in: query
          required: false
          schema:
            type: string
        - name: tipus_tramit
          in: query
          required: false
          schema:
            type: string
        - name: qualificaci_emissions
          in: query
          required: false
          schema:
            type: string
        - name: emissions_calefacci
          in: query
          required: false
          schema:
            type: number
        - name: qualificaci_emissions_1
          in: query
          required: false
          schema:
            type: string
        - name: emissions_refrigeraci
          in: query
          required: false
          schema:
            type: number
        - name: qualificaci_emissions_acs
          in: query
          required: false
          schema:
            type: string
        - name: emissions_acs
          in: query
          required: false
          schema:
            type: number
        - name: emissions_enllumenament
          in: query
          required: false
          schema:
            type: number
        - name: qualificaci_energia_calefacci
          in: query
          required: false
          schema:
            type: string
        - name: energia_calefacci
          in: query
          required: false
          schema:
            type: number
        - name: qualificaci_energia
          in: query
          required: false
          schema:
            type: string
        - name: energia_refrigeraci
          in: query
          required: false
          schema:
            type: number
        - name: qualificaci_energia_acs
          in: query
          required: false
          schema:
            type: string
        - name: energia_acs
          in: query
          required: false
          schema:
            type: number
        - name: energia_enllumenament
          in: query
          required: false
          schema:
            type: number
        - name: qualificaci_energia_calefacci_1
          in: query
          required: false
          schema:
            type: string
        - name: energia_calefacci_demanda
          in: query
          required: false
          schema:
            type: number
        - name: qualificaci_energia_2
          in: query
          required: false
          schema:
            type: string
        - name: energia_refrigeraci_demanda
          in: query
          required: false
          schema:
            type: number
        - name: longitud
          in: query
          required: false
          schema:
            type: number
        - name: latitud
          in: query
          required: false
          schema:
            type: number
        - name: rehabilitacio_energetica
          in: query
          required: false
          schema:
            type: string
        - name: data_entrada
          in: query
          required: false
          schema:
            type: string
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DadesObertesEdifici'
        400:
          description: 'Invalid parameters in query'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error400'
  /dadesObertes/alternatives:
    get:
      tags:
        - Dades Obertes
      summary: Returns all results from DadesObertes of buildings within a radius centered at indicated point.
      operationId: getAlternativesDadesObertes
      parameters:
        - name: latitude
          in: query
          required: false
          schema:
            type: number
            example: 41.388099
        - name: longitude
          in: query
          required: false
          schema:
            type: number
            example: 2.148293
        - name: radius
          in: query
          required: false
          schema:
            type: number
            example: 100
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DadesObertesEdifici'
        400:
          description: 'Invalid parameters in query'
          content:
            application/json:
              examples:
                errorExample:
                  $ref: '#/components/examples/error400'
...
components:
  schemas:
...
    DadesObertesEdifici:
      type: object
      properties:
        num_cas:
          type: string
        adre_a:
          type: string
        numero:
          type: string
        escala:
          type: string
        pis:
          type: string
        porta:
          type: string
        codi_postal:
          type: string
        poblacio:
          type: string
        comarca:
          type: string
        nom_provincia:
          type: string
        codi_poblacio:
          type: string
        codi_comarca:
          type: string
        codi_provincia:
          type: string
        referencia_cadastral:
          type: string
        zona_climatica:
          type: string
        metres_cadastre:
          type: number
        us_edifici:
          type: string
        qualificaci_de_consum_d:
          type: string
        energia_prim_ria_no_renovable:
          type: number
        qualificacio_d_emissions:
          type: string
        emissions_de_co2:
          type: number
        consum_d_energia_final:
          type: number
        cost_anual_aproximat_d_energia:
          type: number
        vehicle_electric:
          type: string
        solar_termica:
          type: string
        solar_fotovoltaica:
          type: string
        sistema_biomassa:
          type: string
        xarxa_districte:
          type: string
        energia_geotermica:
          type: string
        eina_de_certificacio:
          type: string
        valor_aillaments:
          type: number
        valor_finestres:
          type: number
        motiu_de_la_certificacio:
          type: string
        valor_aillaments_cte:
          type: number
        valor_finestres_cte:
          type: number
        utm_x:
          type: number
        utm_y:
          type: number
        normativa_construcci:
          type: string
        tipus_tramit:
          type: string
        qualificaci_emissions:
          type: string
        emissions_calefacci:
          type: number
        qualificaci_emissions_1:
          type: string
        emissions_refrigeraci:
          type: number
        qualificaci_emissions_acs:
          type: string
        emissions_acs:
          type: number
        emissions_enllumenament:
          type: number
        qualificaci_energia_calefacci:
          type: string
        energia_calefacci:
          type: number
        qualificaci_energia:
          type: string
        energia_refrigeraci:
          type: number
        qualificaci_energia_acs:
          type: string
        energia_acs:
          type: number
        energia_enllumenament:
          type: number
        qualificaci_energia_calefacci_1:
          type: string
        energia_calefacci_demanda:
          type: number
        qualificaci_energia_2:
          type: string
        energia_refrigeraci_demanda:
          type: number
        longitud:
          type: number
        latitud:
          type: number
        georeferencia:
          type: object
          properties:
            type:
              type: string
            coordinates:
              type: array
              items:
                type: number
        rehabilitacio_energetica:
          type: string
        data_entrada:
          type: string
...
  examples:
...
    error400:
      value:
        message: 'Invalid information entered'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT